+++ filein.txt
@@ -1,8 +1,4 @@
-2
-12 18
-2
-24 5
 1
-12 169
-1
-12 -169
+4 5
+2
+9 3
